backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
scikit-learn>=1.7.0
joblib>=1.3.2
numpy>=1.26.0,<2.0
pandas==2.1.3
email-validator==2.1.0
google-generativeai>=0.3.0
pillow>=10.0.0

